优化sharpness着色器代码

+++ libausbc/src/main/res/raw/effect_sharpness_fragment.glsl
@@ -13,11 +13,11 @@ uniform sampler2D uTextureSampler;
 
 void main()
 {
-     mediump vec3 textureColor = texture2D(uTextureSampler, vTextureCoord).rgb;
-     mediump vec3 leftTextureColor = texture2D(uTextureSampler, leftTextureCoordinate).rgb;
-     mediump vec3 rightTextureColor = texture2D(uTextureSampler, rightTextureCoordinate).rgb;
-     mediump vec3 topTextureColor = texture2D(uTextureSampler, topTextureCoordinate).rgb;
-     mediump vec3 bottomTextureColor = texture2D(uTextureSampler, bottomTextureCoordinate).rgb;
+     highp vec3 textureColor = texture2D(uTextureSampler, vTextureCoord).rgb;
+     highp vec3 leftTextureColor = texture2D(uTextureSampler, leftTextureCoordinate).rgb;
+     highp vec3 rightTextureColor = texture2D(uTextureSampler, rightTextureCoordinate).rgb;
+     highp vec3 topTextureColor = texture2D(uTextureSampler, topTextureCoordinate).rgb;
+     highp vec3 bottomTextureColor = texture2D(uTextureSampler, bottomTextureCoordinate).rgb;
 
      gl_FragColor = vec4((textureColor * centerMultiplier -(leftTextureColor * edgeMultiplier + rightTextureColor * edgeMultiplier + topTextureColor * edgeMultiplier + bottomTextureColor * edgeMultiplier)), texture2D(uTextureSampler, bottomTextureCoordinate).w);
 }
